OCT-111: moved SearchFamilyQuery
