+++ components/catalogs/back/src/Infrastructure/Controller/Internal/GetFamiliesAction.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Akeneo\Catalogs\Infrastructure\Controller\Internal;
 
 use Akeneo\Catalogs\Application\Persistence\Family\GetFamiliesByCodeQueryInterface;
-use Akeneo\Catalogs\Application\Persistence\SearchFamilyQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Family\SearchFamilyQueryInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
