+++ components/catalogs/back/src/Infrastructure/Symfony/Resources/config/services.yml
@@ -68,7 +68,7 @@ services:
         arguments:
             $limit: '%akeneo_catalog.max_number_of_catalogs_per_user%'
 
-    Akeneo\Catalogs\Infrastructure\Persistence\SearchFamilyQuery:
+    Akeneo\Catalogs\Infrastructure\Persistence\Family\SearchFamilyQuery:
         arguments:
             - '@pim_enrich.repository.family.search'
