+++ components/catalogs/back/tests/Unit/Infrastructure/Controller/Internal/GetFamiliesActionTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Akeneo\Catalogs\Test\Unit\Infrastructure\Controller\Internal;
 
 use Akeneo\Catalogs\Application\Persistence\Family\GetFamiliesByCodeQueryInterface;
-use Akeneo\Catalogs\Application\Persistence\SearchFamilyQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Family\SearchFamilyQueryInterface;
 use Akeneo\Catalogs\Infrastructure\Controller\Internal\GetFamiliesAction;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\RedirectResponse;
